Rename bisq2 intro to welcome
