+++ desktop/src/main/java/bisq/desktop/common/view/NavigationTarget.java
@@ -41,7 +41,7 @@ public enum NavigationTarget {
     TAC(OVERLAY, false),
 
     ONBOARDING(OVERLAY, false),
-    ONBOARDING_BISQ_2_INTRO(ONBOARDING, false),
+    ONBOARDING_WELCOME(ONBOARDING, false),
     ONBOARDING_GENERATE_NYM(ONBOARDING, false),
     ONBOARDING_PASSWORD(ONBOARDING, false),
